Move Java TSP algorithm to org.coiner package
