+++ src/main/java/org/coinor/tsp/MyGreedyStartSolution.java
@@ -1,4 +1,4 @@
-package tspjava;
+package org.coinor.tsp;
 
 
 public class MyGreedyStartSolution extends MySolution
